STP-949 Schmervice causing issues with multiple sever starts
